+++ src/Behat/Context/Domain/Proxy/PackageContextProxyInterface.php
@@ -1,13 +1,5 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
-
 namespace Proximum\Vimeet\Behat\Context\Domain\Proxy;
 
 use Proximum\Vimeet\Behat\Context\Storage\StorageInterface;
@@ -15,13 +7,6 @@ use Proximum\Vimeet\Behat\Service\Manager\PackageManager;
 
 interface PackageContextProxyInterface
 {
-    /**
-     * @return StorageInterface
-     */
     public function getStorage(): StorageInterface;
-
-    /**
-     * @return PackageManager
-     */
     public function getPackageManager(): PackageManager;
 }