+++ src/Infrastructure/Adapter/UuidGenerator.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Proximum\Vimeet\Infrastructure\Adapter;
+
+use Proximum\Vimeet\Application\Adapter\UuidGeneratorInterface;
+use Ramsey\Uuid\Uuid;
+
+class UuidGenerator implements UuidGeneratorInterface
+{
+    public function generate(): string
+    {
+        return Uuid::uuid4()->toString();
+    }
+}