+++ CHANGELOG.md
@@ -4,9 +4,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
-- VIMEET-2173 - Add option submit validation sheet
 - VIMEET-2106 - The speaking user video is maximized based on audio level
+
+## [2.43.0] - 2020-08-27
+### Added
+- VIMEET-2187 - Invisible mode for webinar speaker
+- VIMEET-2174 - Do not ask a new password when user is logged with a token
+
+## [2.42.0] - 2020-07-31
+### Added
+- VIMEET-2168 - Video object on sheet template
+- Add fly system and google cloud storage adapter
+- VIMEET-2173 - Add option submit validation sheet
+- VIMEET-2170 - Upload Video on sheet
 
 ## [2.41.0] - 2020-07-30
 ### Fixed